feat: manual GG/anno override per profile in Business Plan
- Add getEffectiveDaysYear helper in teamUtils.js
- TeamCompositionTable: editable GG/anno input with amber indicator for manual overrides
- businessPlanEngine: per-member effective daysPerFte respects manual_days_year
- ProfileMappingEditor: cost calculations use effective days per member
- TowAnalysis: per-member contribution cost uses effective days
- FTE change auto-resets manual override
- i18n: resource modal + delegated resources translations
- Fix: framer-motion import in TechEvaluator
- Fix: prof_certs_resources persistence in backend seed

+++ backend/services/scoring_service.py
@@ -68,11 +68,11 @@ class ScoringService:
 
     @staticmethod
     def calculate_professional_score(
-        resources: int,
-        certifications: int,
-        max_resources: int,
+        resources: float,
+        certifications: float,
+        max_resources: float,
         max_points: float,
-        max_certifications: int = 5
+        max_certifications: float = 5.0
     ) -> float:
         """
         Calculate professional score based on resources and certifications.
